+++ lib/src/dart/throttler.dart
@@ -7,7 +7,18 @@ class Throttler {
 
   Throttler({required this.milliseconds});
 
+  @Deprecated("Use call instead")
   void run(Function action) {
+    if (!_isReady) return;
+    final duration = Duration(milliseconds: milliseconds);
+    _isReady = false;
+    action();
+    _timer = Timer(duration, () {
+      _isReady = true;
+    });
+  }
+
+  void call(Function action) {
     if (!_isReady) return;
     final duration = Duration(milliseconds: milliseconds);
     _isReady = false;